Create a utility to add/remove entries from a link list within a given buffer
This buffer is designed to be used in a runtime environment.

I am laying the ground work for an over arching story that will
save traces in a buffer for future consumption.

For this commit, I have created the methods necessary to add, remove and find
available space (for an Entry) in a buffer bound region of memory using a
circular, doubly linked list.

Change-Id: I259bb0f6051611a17b7b919bf026919ffdb12eb1
RTC: 191303
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/59575
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/60919