Link PLID for failing SBE recovery in PRD path w/ other related logs
When Hostboot fails to recover the SBE after PRD has alerted HB that
the SBE needs recovery then HB will create an error log. Prior to this
commit this log was not getting it's PLID linked with the other logs
that get generated over the course of the sbe recovery attempts. This
commit ensures that if any logs were created during the sbe recovery
that they will be linked with the final log that tells us that the
recovery was unsuccessful.

Change-Id: I7af47cf437af0b68b954ab2b0e95aec660c11bab
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/59648
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>