Look for any parent on deconfigure
There are error scenarios where a unit can get deconfigured
and its parent is not considered present. This can lead to errors
in the deconfig by association logic because a unit should always
have a present parent.

This change has the deconfig logic just find the parent. If it's
considered not present or not functional then the code can handle
that, it just needs the parent target.

This commit also fixes a bug found in testing in the DIMM deconfig
by association path.

Change-Id: I98c4185192a0944543cc545277504b50bd951c3e
CQ: SW431931
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/60190
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>