Use last l3 cache object for SIMICS trace
The logic used in this script to choose the correct l3_cache_ram
object is complex. Currently in a multi-node SIMICS system,
the first l3_cache_ram object is always used because it always
has an HRMOR that is at 0x8000000 and all l3_cache_ram objects
have the same priority (0). This change just adjusts the logic
to always use the last l3_cache_ram object in the list that
fits the required memory region and has the same or higher
priority.

Change-Id: I960a5b9f04d934a8823f541c0a95133700cc0120
CQ: SW431167
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/60050
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>