New Global For Console Daemon
This change implements a new global variable that indicates
whether the console daemon has been started. This variable
is checked in the doShutdown path prior to printing any
console traces.

There may be a situation very early in the IPL where we
attempt to write to console before the console daemon has
been started. This causes timeouts and IPL failures.

Change-Id: Ibec90d2207e5db9ef3b9507503d1ea9e2eb7cf6d
CQ: SW431130
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/60227
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>