Fix for multinode HBRT use of VPD
Node0 VPD memory is being used for all nodes because we were
using a Singleton pointer across modules.  This is a problem
because static memory is specific to each module which meant
we were getting a new copy of the AttrRP object instead of
the real version that had real data in it.

The change here is to create a new static interface that
external modules can call to retrieve the node associated
with a given Target.

There is also a change to cache this data into a map as I
noticed hundreds of accesses during the one function call
I was using to test with.

Change-Id: I148bf1f405d076276193d526d8a4b1f0649b2c1c
CQ: SW431462
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/60266
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>