Fix bugs in core checkstop escalation manipulation during HB
There was a bug where in fused mode we are not also setting the 2nd
half of the fused core to have the appropriate scom values in the fir
action registers. Also we were not re-applying the desired values back
to the core fir action registers at the end of hostboot because of a
simple bug.

Change-Id: I32ae9c7839e27392cfdf0401d008a12941804beb
CQ: SW432725
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/60614
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>