Ensure hwas state reflects resource recovery actions
Once resource recovery is run to recover hardware resources to
allow the system to boot, it will not be run again until the
configuration of the hardware changes. An issue with that design
is that the HWAS state, deconfiguredByEid, will not be properly
updated in situations where resource recovery has already run.

This state is used by the gard command line tool to inform
the user on the state of their garded hardware.

This commit ensure the deconfiguredByEid is properly updated
for targets that are guarded, but have been resource recovered
to allow system boot.

Change-Id: Ib15c1e8402e7c13b6497915c5138831e5e591bbe
CQ: SW432846
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/60883
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Disable-CI: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>