getMBvpdAttr: Updates for DDR3 support
Changed the check from if (rc == false) to if (rc != success).
This way we are not relying on plat code to return a specific
rc in case VPD rec/kw is not found. This HWP relies on a few
VPD rec/kw combinations that will not work and we need to
ignore errors in that case. The HWP tries the combinations in
the following order: VSPD+VM, VSPD+VD, VINI+VZ. For DDR3, first
two combinations do not work.

Change-Id: I43966493a1d7e9e2f6d9447dd3fda9c3206de417
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/60507
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/60508
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>