Pre-set HB TI Area in doStutdown Path
When a TI occurs during ECC corruption, the kernel asserts as part
of the error flow, and the assert makes it into the TI area before
the ECC error code, which makes it hard to debug the underlying
issue. This change introduces the logging of the TI code to the
HB TI area as part of the doShutdown path and a mechanism to not
overwrite the first logged error code. That way the TI area will
always contain the first error code that caused the TI.

Change-Id: Idcd5727314aea9b92a6eb9d19ec6ed223111861a
CQ:SW431570
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/61375
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>