Allow HWSV to handle gard callouts during runtime for FSP systems
There was a bug in the runtime code where if a errorlog was created
and it had a deconfigure callout as well as a gard callout it would
attempt to create a gard record for the callout. This caused PNOR
reads which are not permitted in HBRT on systems with FSPs.

Change-Id: Ie2e1019889e0087396acfa07c7ce93e24643f322
CQ: SW430100
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/61505
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>