Always use last valid SIMICS object for trace
The script should always use the last SIMICS object it finds
that covers the required memory area.

This fix was required because there is a scenario where there
are 2 nodes booting in parallel in an SMP based system. node1
reached istep15 and exited cache contained mode which removed
its D2Proc0.l3_cache_ram object. node0 however was still in
cache contained mode. The cache objects in SIMICS have a priority
of 0 (the highest) and the memory objects have a priority of 1.
Therefore when we hit this situation, the SIMICS object for
D1Proc0.l3_cache_ram was used incorrectly for node1 memory.

The fact this script uses node0 by default, and then looks for
a higher priority object that covers the correct memory region
is confusing. But, there are historical reasons it had to be
done that I'd rather not revisit.

Change-Id: I46521b6252c25f3ca77022c41e0035a065ca2bfb
CQ:SW435477
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/61511
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>