Only switch sides and perform hreset if SEEPROM side versions match
In the event that the SBE fails hostboot will attempt to recover it.
During runtime hostboot will attempt an HRESET if the SBE is in a
failed state. When the SBE performs the HRESET it will save some
important information that will persist through the reset. If one
side is failing to recover the retry code will attempt to switch sides
and do the hreset. If the SBE seeproms have different versions of the
SBE code the data that was supposed to persist through the HRESET will
be in incorrect places because the version mismatch. Because of this
we cannot switch seeprom sides and perform a hreset if the seeproms have
different level of the SBE code.

CQ: SW438029
Change-Id: Ic7078a886088cc4d5355cc076e72d0fc36f85027
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/61605
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>