Adding VPD testing data for Cumulus and Nimbus.
We turned off some VPD tests when we branched to include
cumulus. However, the VPD is different between the two and the
tests have to be different to match. This commit splits up the
testing array into two and tests depending on what system we're on.
I picked out a random sampling of VPD keyword record pairs.

Change-Id: I56949b260278e132008cc9c22068a41cc9f3bc7d
RTC:190805
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/61909
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>