PRD: ignore SCOM errors in rule code summary construct
This is a bit unusual, but we are going to ignore SiCOM failures.
This takes care of a corner case where one of the summary registers
in the list returns an error, but there is another summary register
with an active attention, which would be ignored if we gave a bad
return code.

Change-Id: I25ae90ac424e28422aab5d48af128b03ccfb1344
CQ: SW436299
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/62083
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Brian J. Stegmiller <bjs@us.ibm.com>
Reviewed-by: Caleb N. Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Benjamin J. Weisenbeck <bweisenb@us.ibm.com>
Reviewed-by: Zane C. Shelley <zshelle@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/62147
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/62280
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>