Only store HRMOR value on primary node
There is a single field in HDAT where Hostboot fills in
an address where PHYP will then write the new HRMOR
value any time it changes.  The problem is that the
independent Hostboot instance in each node is writing
the same field with their own node-relative address.
There is only 1 field so the last node in will win.
Then on the mpipl we won't be able to find it.

Change-Id: I18f64816f20b0a3fc2951dff457d9511ca27e1d2
CQ:SW438101
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/62763
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Richard Ward <rward15@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>