Add hostboot reserved memory mirroring support
Adding this support to avoid losing runtime functions
(PRD, OCC/PM Complex management) to a mirroring-enabled
system when a memory channel failure happens to the
first piece of memory in a drawer.

Basically just need to change where we physically map
hostboot reserved memory to be the new mirrored spot.

Change-Id: I85aa8c6033a47e4598eb1130e0b3d06c360ea479
CQ:SW437433
CMVC-Prereq: 1063550
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/63034
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: ILYA SMIRNOV <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>