Host went to kernel panic after gard'ing PEC0/PHB0 (PCIFIR CS)
The host kernel was crashing when it probed PEC0/PHB0.  HDAT
was populating incorrectly, not showing that PEC0/PHB0 was
gard'd.  The code which populates HDAT uses the attribute,
ATTR_PROC_PCIE_PHB_ACTIVE, to indicate which PHBs are active.
I updated hostboot code to consider gards when setting the
attribute.

Change-Id: Ic1c208af6ed67d7b520c7c16ef1836be1dd66170
CQ:SW412466
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/63295
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>