Port System and Node Targets Stitching Code
HBRT's targeting doesn't map the System target to its
nodes correctly because the target stitching code, that
already exists on FSP, is not present in HBRT. This change ports
the abovementioned code to HBRT, so that the system
and node targets are correctly associated to each other
at the end of targeting init. Note that unstitching is required
during MPIPL to break the links between sys <-> node because
only one node on which hb is running is visible during the boot.

Change-Id: I9731144c480b3357ff7b4aff02d13623a1119dac
CQ: SW432119
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/65494
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>