Support for mrwHide attribute from fapi attribute xmls
We are adding a new tag to the fapi xml files that allows the
HWP team to indicate that a platInit attribute should be hidden
from ServerWiz2.  This tag is converted to our existing no_export
tag in targeting that is consumed during our push out to the
serverwiz repo.

Change-Id: I7311c81e6538956f03365f8b3ea2ce93cedec2ff
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/63819
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>