WOF Lid load: Remove extra malloc and use getStoredLidImage
Currently, both hbrt and phyp are calling malloc when loading
WOF lids. This causes heap overflow when the heap is super
fragmented. So, a quick way to avoid heap overflow is to
remove malloc from HBRT. Additionally, we want to use getStoredLidImage
because it will return the cached copy of the lid if it exists.

Change-Id: I9bf6e8ec7c94d97a1998a56482f617608835d635
CQ:SW441541
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/64270
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Richard Ward <rward15@us.ibm.com>
Reviewed-by: ILYA SMIRNOV <ismirno@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>