lpc_init: Correct LPC host controller timeout value
The LPC host controller has an interesting way to decode the timeout
value. The left 4 bits are used for the "short wait" timeout, while
the entire 8 bits are used for the "long wait" timeout. If the "short
wait" timeout is 0xF, it is taken to be infinite, causing the host
controller to hang if the slave doesn't respond.

Change the timeout value from 0xFE to 0xEF, the correct maximum value
that is not decoded to be infinity.

Change-Id: Iaf1a5119a87338c24b1e324d814ade0b30353360
CQ: SW442999
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/64850
Reviewed-by: Joseph J. McGill <jmcgill@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>