Trusted Boot: Report Primary TPM required reason code as terminating
This commit flags the RC_TPM_NOFUNCTIONALTPM_FAIL reason code as a terminating
reason code to tell FSP that there is an expected boot failure without a
deconfiguration.  Further, when this error occurs, it TIs with this error's
reason code instead of its PLID, so that FSP will balk to the alignment check to
failover and use the backup TPM, when possible.

Change-Id: I6809dd1de6f37ce50d6528ce1014c5ede115b0dd
CQ: SW442904
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/64973
Reviewed-by: ILYA SMIRNOV <ismirno@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>