Check capability bit before using wakeup for OPAL in OpenPOWER
A recent change switched out special wakeup handling over to use
the host interface instead of the internal HWP calls.  However,
we just discovered that the opal-prd application was not allowing
those calls on non-FSP systems.  The opal-prd code is being
updated, but since that is delivered as part of the Linux distro
(not firmware) we need to handle the situation where newer
Hostboot is running on top of the older opal-prd.

This commit adds a check for a new OPAL-specific capability bit
that will ensure we only use the wakeup interface if the opal-prd
hosting us supports it.

Change-Id: Ib9b99a4549a023d466572c19c2f7208e1c26d8f8
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/65208
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>