Add OCMB_CHIP and MEM_PORT fapi2 test cases
This commit adds test cases for new OCMB_CHIP and MEM_PORT target
types that are being added to support the new explorer memory buffer
used in Axone.
Also this commit fixed up the Axone target generation which was
missing a break on the inner loop and removed some changes that
were made in target.H which was allowing invalid parets to be
looked up for dimm targets.

RTC: 172970
Change-Id: I7564fbcfcdde2177ece934015fe40c79011f559f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/65511
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>