Forcibly clear all previous wakeups when the PM Complex starts
The power management logic was designed with the assumption
that the external callers would manage the state machine such
that the order of operations is always:
  load->start->reset->load->start->reset->etc.  However, we
have discovered edge cases where a reset can happen before the
first load or after a previous reset.  In either case we will
end up with our wakeup counter being off by one and thus never
releasing wakeup.

The solution is to add logic to the start path that will
explicitly clear the counter out completely, regardless of its
current value.  This will ensure that wakeup is released once
the complex is alive.

Change-Id: Ifb436335090ff910b9ee0f083d9a550b81833287
CQ: SW442639
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/65518
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>