Don't Process Interrupts During Shutdown
An issue has come up on OpenPOWER where we would checkstop during
istep 10.2 or 16.1 when completing the interrupt processing due to
the code attempting to access a memory area that got locked during
the shutdown sequence. This change no-ops the complete interrupt
routine if a shutdown has been requested to remedy this issue.

Change-Id: Ie444149884c6a42a0010593fb94b91298b4bc798
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/66049
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>