Ensure we collect PPE trace if psu op times out
There was a bug in the code where if we are running on a FSP system
in simics we will not run the magic instruction to capture PPE trace
at the time of the psu op timeout. This commit insures we will call
sbe-trace and save the simics console output to file for later debug
this will help with the defect listed in this commit.

Change-Id: I7caa776f06197f7732a5dcabdcc01a3b91cdfba3
CQ: SW444734
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/65992
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>