Trusted Boot: When TPM not detected properly, report errors if TPM is required
When the MRW for a machine defines a possible enabled TPM, and the "TPM
Required" policy is "Required", then it is expected the TPM will be detected as
present. If that is not the case, surface the detection failure as a visible
error log, and generate another visible log indicating that a TPM expected to be
present was not detected correctly.

Change-Id: Icbfca04beaa6fdece147cc24f4e0615b5856a830
CQ: SW444532
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/66422
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>