Correctly handle psu FFDC on OpenPower Systems
Prior to this change, there was a switch statement that was saying
"if this RC is found in the psu FFDC, then do this recovery action"
that is obviously not very easy to maintain because for every error
we need to add the proper action. Instead of this, now we will just
look if any GARD records were created as part of the error found in
the FFDC. If a gard was found , Hostboot will stop trying to recover
the SBE and instead enter a reconfig loop to try and IPL w/ the target
garded out. Again this only applies to OP system, in the FSP world we
will commit the error logs w/ the gard records and then TI telling HWSV
they need to look at the SBE

Change-Id: I04e03feebf2bbd1eae2d725bee31993062fe7c94
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/66374
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>